api/requirements.txt
fastapi==0.110.0
openai==1.65.2
python-dotenv==1.0.0
loguru==0.7.2
pydantic==1.10.12
requests==2.32.3
uvicorn==0.24.0